+++ src/highlight/index.ts
@@ -1,3 +1,5 @@
 export * from "./HighlightPresenter";
+export * from "./jsonTokenRanges";
+export * from "./syntaxHighlightConfig";
 export * from "./TextDecorator";
 export * from "./types";