+++ src/highlight/types.ts
@@ -1,4 +1,5 @@
 import * as vscode from "vscode";
+import type { JsonTokenKind } from "./jsonTokenRanges";
 
 export type TextRange = {
   range: vscode.Range;
@@ -16,4 +17,15 @@ export const defaultFragmentDecorationStyle: FragmentDecorationStyle = {
   backgroundColor: new vscode.ThemeColor("editor.findMatchHighlightBackground"),
   overviewRulerColor: new vscode.ThemeColor("editorOverviewRuler.findMatchForeground"),
   borderRadius: "2px",
+};
+
+export type JsonTokenDecorationStyle = {
+  color: vscode.ThemeColor;
+};
+
+export type JsonTokenDecorationStyles = Record<JsonTokenKind, JsonTokenDecorationStyle>;
+
+export type HighlightRenderLayers = {
+  fragment: boolean;
+  inlineSyntax: boolean;
 };